move WalletShieldedOutput
